Add files via upload
The purpose of this study is: (1) to determine the validity level of the electric angle clock learning media, (2) to determine the level of attractiveness of the electric angle clock learning media, (3) to determine the effectiveness level of electric angle clock learning media for class IV SD Negeri Pangeranan 3 Bangkalan. This study uses the Borg and Gall development model which was modified by Sugiyono. Based on the results of the research conducted, it was obtained the results that the validity percentage of the electric angle clock learning media according to the instructional media expert was 91.25% (very valid), the learning design expert was 76% (valid), and the material expert was 83% (valid). The attraction of the electric angle clock learning media was measured by the percentage of student response questionnaire results in the usage test of 88.4% (very interesting). The effectiveness of instructional media is measured from the percentage of results in the use trial including the percentage of observations of teacher activity by 93.9% (very active), the percentage of observations of student activity by 82.9% (active), and the percentage of classical learning completeness by 86.2 % (complete). These results indicate that the electric angle clock learning media is valid, very interesting, and very effective.